+++ src/main/java/org/yandrut/selenium/manager/ChromeManager.java
@@ -1,0 +1,18 @@
+package org.yandrut.selenium.manager;
+
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
+
+public final class ChromeManager {
+
+    private ChromeManager(){}
+
+    public static WebDriver getDriver() {
+        ChromeOptions options = new ChromeOptions();
+        options.addArguments("start-maximized");
+        options.addArguments("--disable-notifications");
+        options.addArguments("--disable-popup-blocking");
+        return new ChromeDriver(options);
+    }
+}
