+++ src/main/java/org/yandrut/selenium/manager/EdgeManager.java
@@ -1,0 +1,13 @@
+package org.yandrut.selenium.manager;
+
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.edge.EdgeDriver;
+
+public final class EdgeManager {
+
+    private EdgeManager() {}
+
+    public static WebDriver getDriver() {
+        return new EdgeDriver();
+    }
+}
